Fix code-review findings from final review of today's changes
Code review of the full day's diff (v1/v2 pipeline work) surfaced 10
findings. Fixes applied:

- process_v2_batch_item created a duplicate CTExamination for jobs
  queued by Manual Exam Entry (which already creates its own
  CTExamination synchronously). Confirmed live: two rows with the same
  repository_study_id linked to one job. Guarded with
  job.examinations.exists().
- v2 batch items with a declared sha256 now have their archive checksum
  verified before extraction/anonymization, mirroring the v1 pipeline's
  pre-flight integrity check.
- Patient.age_at_first_acquisition was a PositiveIntegerField, silently
  truncating fractional ages (e.g. 0.3) via int(0.3) == 0. Migrated to
  DecimalField, consistent with CTExamination.patient_age.
- Decimal('0') is falsy: patient_age_years in manifest_raw used a
  truthiness check ("if patient_age") instead of "is not None",
  dropping a valid age of 0.
- UploadJobDeleteView could delete tar_temp_path out from under a
  CTExamination that still references the same file for download
  (Manual Exam Entry jobs) -- now only removes genuinely orphaned
  staging archives.
- Missing PseudoIDUniquenessValidator.log_pseudo_id_tracking audit call
  after patient creation in the v2 pipeline.
- Manual Exam Entry no longer queues the async pipeline when
  repository_study_id generation failed, avoiding an empty/malformed
  ID prefix.
- ChunkedUploadInitView now validates manifest_item against the v2
  item schema, closing the gap where posting directly to the endpoint
  bypassed the client-side Validate Manifest step.
- automated_upload.js's batch upload handler always showed a 'success'
  banner regardless of per-item failures; uploadOneItem() now returns
  a real boolean on every path.
- manifest_schema.py's v2 item schema now requires minLength: 1 on
  ref/filename, so an empty string can no longer bypass the duplicate
  detection's truthy checks.

SessionAuthentication alongside BearerTokenAuthentication (added
earlier today for the Automated Upload GUI page) was reviewed as an
intentional, already-tested widening of the DRF auth surface -- not a
bug, no code change needed.

Added regression tests for each fix; full suite: 772/772 passing.
Migration 0036 applied; static files recollected for automated_upload.js.

+++ ct_upload_platform/uploads/pseudo_id_validator.py
@@ -65,7 +65,7 @@ class PseudoIDUniquenessValidator:
     def get_or_create_patient_with_pseudoid(
         pseudo_id: str,
         sex: Optional[str] = None,
-        age_at_acquisition: Optional[int] = None,
+        age_at_acquisition: Optional[float] = None,
         cohort_tag: Optional[str] = None
     ) -> Tuple[object, bool, Optional[str]]:
         """
